lib/pyproject.toml
[project]
name = "problib"
version = "0.1.0"
description = "A Problem solving library (problib) for working with mathematical optimization problems."
readme = "README.md"
authors = [
    { name = "Karl Jurek Blumenthal", email = "karl.blumenthal@student.htw-berlin.de" },
]
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.8.1",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "rich-structlog>=0.4.1",
    "snakeviz>=2.2.2",
]

[project.scripts]
lib = "lib:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
